requirements.txt
flask==3.0.2
pandas==2.2.1
openai
python-dotenv==1.0.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
fpdf==1.7.2
scikit-learn==1.4.2
statsmodels==0.14.1